requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi
charset-normalizer
cmdkit
croniter
discord-py-interactions
discord-typings
emoji
frozenlist
idna
interactions.py
multidict
mysql-connector-python
names-generator
protobuf
python-dateutil
python-Levenshtein
pytz
requests
six
tomli
typing_extensions
urllib3
yarl
